pyproject.toml
[project]
name = "CABench"
version = "0.0.1"
description = ""
authors = [
    { name = "Pham Tung Thuy"},
    { name = "Duong Minh Quan"},
    { name = "Luong Duy Quan"}
]
license = { file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "anthropic==0.57.1",
    "bs4==0.0.2",
    "click==8.2.1",
    "codebleu==0.7.0",
    "diffusers==0.34.0",
    "gdown==5.2.0",
    "loguru==0.7.3",
    "nltk==3.9.1",
    "numpy==2.2.6",
    "openai==1.93.0",
    "pandas==2.3.0",
    "pathlib==1.0.1",
    "Pillow==11.3.0",
    "pydantic==2.11.7",
    "PyYAML==6.0.2",
    "requests==2.32.4",
    "scikit-learn==1.7.0",
    "soundfile==0.13.1",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "torch==2.7.1",
    "tqdm==4.67.1",
    "tree-sitter-python==0.21",
    "tree_sitter==0.22.3",
]

[tool.poetry]
package-mode = false